// src/client/types.gen.ts
export type Room = {
    id: string;
    name: string;
    capacity: number | null;
    equipment: string;
    campus: string;
};